*** empty log message ***

+++ GameTree.hpp
@@ -93,8 +93,8 @@ private:
   value_type minmaxValue;	/**< minmax value  */
   uint8_t minmaxDepth;		/**< minmax depth */
 
-  int8_t x : 4;			/**< x of last placed piece, or -1 */
-  int8_t y : 4;			/**< y of last placed piece, or -1 */
+  int x : 4;			/**< x of last placed piece, or -1 */
+  int y : 4;			/**< y of last placed piece, or -1 */
 
   TreeNode *swapPlayer();
 };
